refactor: decouple IProviderAdress from Mongoose for flexibility
