+++ apps/backend/api/types/ProviderAddress.types.ts
@@ -1,0 +1,15 @@
+interface Location {
+  type: "Point";
+  coordinates: [number, number];
+}
+
+export interface IProviderAddress {
+  code: string;
+  city: string;
+  street: string;
+  building_number: string;
+  district: string;
+  post_code: string;
+  voivodeship: string;
+  location: Location;
+}
